# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DatabaseInfo": {
        "description": "Information about a configured database connection.",
        "properties": {
          "allow_writes": {
            "description": "Whether write operations are permitted on this connection.",
            "type": "boolean"
          },
          "database": {
            "description": "The database name.",
            "type": "string"
          },
          "host": {
            "description": "The database server hostname.",
            "type": "string"
          },
          "name": {
            "description": "The logical name of the database connection.",
            "type": "string"
          },
          "port": {
            "description": "The database server port.",
            "type": "integer"
          },
          "sslmode": {
            "description": "The SSL mode for the connection.",
            "type": "string"
          },
          "user": {
            "description": "The database user.",
            "type": "string"
          }
        },
        "required": [
          "name",
          "host",
          "port",
          "database",
          "user",
          "sslmode",
          "allow_writes"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "A generic error response.",
        "properties": {
          "error": {
            "description": "A human-readable error message.",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "JSONRPCRequest": {
        "description": "A JSON-RPC 2.0 request object for the MCP protocol.",
        "properties": {
          "id": {
            "description": "A unique request identifier. Can be a string or integer."
          },
          "jsonrpc": {
            "description": "The JSON-RPC protocol version. Must be \"2.0\".",
            "enum": [
              "2.0"
            ],
            "type": "string"
          },
          "method": {
            "description": "The MCP method to invoke.",
            "enum": [
              "initialize",
              "tools/list",
              "tools/call",
              "resources/list",
              "resources/read",
              "prompts/list",
              "prompts/get"
            ],
            "type": "string"
          },
          "params": {
            "description": "Optional parameters for the method.",
            "type": "object"
          }
        },
        "required": [
          "jsonrpc",
          "id",
          "method"
        ],
        "type": "object"
      },
      "JSONRPCResponse": {
        "description": "A JSON-RPC 2.0 response object.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/RPCError",
            "description": "Error information. Present on failure."
          },
          "id": {
            "description": "The request identifier that this response corresponds to."
          },
          "jsonrpc": {
            "description": "The JSON-RPC protocol version.",
            "enum": [
              "2.0"
            ],
            "type": "string"
          },
          "result": {
            "description": "The result of the method invocation. Present on success."
          }
        },
        "required": [
          "jsonrpc",
          "id"
        ],
        "type": "object"
      },
...
      "ListDatabasesResponse": {
        "description": "Response containing the list of databases and the currently selected database.",
        "properties": {
          "current": {
            "description": "The name of the currently selected database.",
            "type": "string"
          },
          "databases": {
            "description": "The available database connections.",
            "items": {
              "$ref": "#/components/schemas/DatabaseInfo"
            },
            "type": "array"
          }
        },
        "required": [
          "databases",
          "current"
        ],
        "type": "object"
      },
...
      "RPCError": {
        "description": "A JSON-RPC 2.0 error object.",
        "properties": {
          "code": {
            "description": "A numeric error code.",
            "type": "integer"
          },
          "data": {
            "description": "Additional error data."
          },
          "message": {
            "description": "A short description of the error.",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "SelectDatabaseRequest": {
        "description": "Request to select a database by name.",
        "properties": {
          "name": {
            "description": "The name of the database to select.",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "SelectDatabaseResponse": {
        "description": "Response after selecting a database.",
        "properties": {
          "current": {
            "description": "The name of the now-current database.",
            "type": "string"
          },
          "error": {
            "description": "An error message if the operation failed.",
            "type": "string"
          },
          "message": {
            "description": "A human-readable status message.",
            "type": "string"
          },
          "success": {
            "description": "Whether the database was selected successfully.",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "current",
          "message"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "token",
        "description": "Bearer token authentication. Required when auth is enabled on the server.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "pgEdge",
      "url": "https://www.pgedge.com"
    },
    "description": "REST API for the pgEdge Postgres MCP Server, providing database management, MCP protocol access, LLM proxy, chat compaction, and conversation management.",
    "license": {
      "name": "PostgreSQL License"
    },
    "title": "pgEdge Postgres MCP Server API",
    "version": "1.0.0-beta3"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/databases": {
      "get": {
        "description": "Returns the list of configured database connections and the currently selected database.",
        "operationId": "listDatabases",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListDatabasesResponse"
                }
              }
            },
            "description": "A list of available databases."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List available databases",
        "tags": [
          "Databases"
        ]
      }
    },
    "/api/databases/select": {
      "post": {
        "description": "Switches the active database connection to the specified database name.",
        "operationId": "selectDatabase",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectDatabaseRequest"
              }
            }
          },
          "description": "The name of the database to select.",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectDatabaseResponse"
                }
              }
            },
            "description": "Database selected successfully."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request body."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Access to the requested database is forbidden."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database not found."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Select a database",
        "tags": [
          "Databases"
        ]
      }
    },
...
    "/mcp/v1": {
      "post": {
        "description": "Accepts a JSON-RPC 2.0 request implementing the Model Context Protocol. Supported methods include initialize, tools/list, tools/call, resources/list, resources/read, prompts/list, and prompts/get. Bearer token authentication is required when auth is enabled.",
        "operationId": "postMCP",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JSONRPCRequest"
              }
            }
          },
          "description": "A JSON-RPC 2.0 request object.",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JSONRPCResponse"
                }
              }
            },
            "description": "JSON-RPC 2.0 response."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Send an MCP JSON-RPC 2.0 request",
        "tags": [
          "MCP"
        ]
      }
    }
  },
...
}